# src/main/resources/template/swagger-input.yml
...
openapi: 3.0.1
info:
  title: OpenApi Specification - Nguyen Test ClickAndLearn
  description: OpenApi documentation for Spring Security
  termsOfService: https://cmapi.com
  contact:
    name: ClickAndLearnProject
    url: https://cmapi.com
    email: contact@cmproj.com
  license:
    name: Licences
    url: https://some-url.com
  version: "1.0"
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  "/category":
    get:
      tags:
        - Category Endpoints
      summary: "[ANORNYMOUS] - GET All Categories"
      operationId: getAllCategories
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CategoryDto"
    post:
      tags:
        - Category Endpoints
      summary: "[ADMIN, MANAGER] - Create Category"
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/BaseDto"
    put:
      tags:
        - Category Endpoints
      summary: "[ADMIN, MANAGER] - Update Category"
      operationId: updateCategory
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoryDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/BaseDto"
    delete:
      tags:
        - Category Endpoints
      summary: "[ADMIN, MANAGER] - Delete Category"
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/BaseDto"
  /category/{categoryId}:
    get:
      tags:
        - Category Endpoints
      summary: "[ANORNYMOUS] - GET Category By Id"
      operationId: getCategoryById
      parameters:
        - in: path
          name: categoryId
          required: true
          description: Categoty ID
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                $ref: "#/components/schemas/CategoryDto"
components:
  schemas:
...
    BaseDto:
      type: object
      properties:
        antType:
          type: string
          enum:
            - success
            - error
            - warning
        message:
          type: string
...
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        slug:
          type: string
        image:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Nguyen Test
      scheme: bearer
      bearerFormat: JWT